Add merge conflict filter
