+++ src/main/util/MergeHelper.groovy
@@ -1,7 +1,13 @@
+package main.util
+
+import main.project.*
+import java.util.regex.Pattern
+import java.util.regex.Matcher
+
 class MergeHelper {
     private static String CONFLICT_INDICATOR = "(CONFLICT)|(CONFLITO)"
 
-    public boolean hasMergeConflict (Project project, MergeCommit mergeCommit) {
+    static public boolean hasMergeConflict (Project project, MergeCommit mergeCommit) {
         Process mergeSimulation = replayMergeScenario(project, mergeCommit)
         String mergeMessage = mergeSimulation.getText()
         
@@ -16,14 +22,14 @@ class MergeHelper {
         return result
     }
 
-    public Process replayMergeScenario(Project project, MergeCommit mergeCommit) {
+    static public Process replayMergeScenario(Project project, MergeCommit mergeCommit) {
         Process checkoutLeft = ProcessRunner.runProcess(project.getPath(), 'git', 'checkout', mergeCommit.getLeftSHA())
         checkoutLeft.waitFor()
 
         return ProcessRunner.runProcess(project.getPath(), 'git', 'merge', mergeCommit.getRightSHA())   
     }
 
-    public Process returnToMaster(Project project) {
+    static public Process returnToMaster(Project project) {
         Process resetChanges = ProcessRunner.runProcess(project.getPath(), 'git', 'reset', '--hard')
         resetChanges.waitFor()
