+++ src/test/TestSuite.groovy
@@ -17,7 +17,7 @@ import main.util.FileManager
 import test.*
 
 @RunWith(Suite.class)
-@SuiteClasses([DeletionTest.class, SameLineTest.class, MergeConflictCountTest.class])
+@SuiteClasses([DeletionTest.class, SameLineTest.class, MergeConflictFilterTest.class])
 public class TestSuite {
     public static Map<String, String> modifiedLines;
